p1

+++ src/main/java/server/project_module05/model/dto/response/shopping_cart/ShoppingCartResponse.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 @Data
 @Builder
 public class ShoppingCartResponse {
-    private Long shoppingCartId;
+    private Integer shoppingCartId;
     private String productName;
     private BigDecimal unitPrice;
     private Integer orderQuantity;
